Refactor tls so it's not pulled into command line utilities unecessarily.  Also change default root and make it overridable.
