+++ atrium/vestibulum/trccarrier/carrierfactory/servercapauth/capauth.go
@@ -10,6 +10,7 @@ import (
 	"github.com/trimble-oss/tierceron-hat/cap"
 	"github.com/trimble-oss/tierceron-hat/cap/tap"
 	"github.com/trimble-oss/tierceron/pkg/capauth"
+	"github.com/trimble-oss/tierceron/pkg/tls"
 	"github.com/trimble-oss/tierceron/pkg/vaulthelper/kv"
 	"google.golang.org/grpc"
 )
@@ -137,7 +138,7 @@ func Memorize(memorizeFields map[string]interface{}, logger *log.Logger) {
 func Start(featherAuth *FeatherAuth, env string, logger *log.Logger) error {
 	logger.Println("Cap server.")
 
-	creds, credErr := capauth.GetServerCredentials(logger)
+	creds, credErr := tls.GetServerCredentials(logger)
 	if credErr != nil {
 		logger.Printf("Couldn't server creds: %v\n", creds)
 		return credErr
